% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Es_from_vitals.R
\name{create_Es_from_vitals}
\alias{create_Es_from_vitals}
\title{Make an Es tibble from the vitals}
\usage{
create_Es_from_vitals(vitals = example_vitals(), T = 15)
}
\arguments{
\item{vitals}{The tibble of vital rates, like that from \code{example_vitals()}.}

\item{T}{The number of years of data to have}
}
\description{
The Es tibble has the number of males and females of each age at each time
and then it also has colums that give the relative age composition each year,
and also the ERRO for each cell from that (using the fecundities).  It also
includes the columns that are used to print the mathematical expressions for
each.
}
