% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_vitals.R
\name{example_vitals}
\alias{example_vitals}
\title{returns a sample set of vital rates}
\usage{
example_vitals()
}
\description{
Super simple.  Primarily here to show how someone
could create a set of vitals in a tibble
}
