% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_S.R
\name{iterate_S}
\alias{iterate_S}
\title{Iterate age structured survival from random recruitment}
\usage{
iterate_S(S, T, N0min = 8000, N0max = 10000, ASZ_size = 500, S_var_denom = 30)
}
\arguments{
\item{S}{the vector of survival probs for surviving
to age 1,...,length(S). The last element will be treated
as 0.}

\item{T}{the number of years to simulate forward.}

\item{N0min}{the mininum N0 each year}

\item{N0max}{the maximum N0 each year.  (it is drawn from a uniform)}

\item{ASZ_size}{when setting up the age structure in the first year, the
values are drawn from a multinomial of this size}

\item{S_var_denom}{random annual variation in survival is drawn from a normal
distribution centered on S with standard deviation S/S_var_denom.  This must be
large enough that you don't get values outside of 0 or 1.}
}
\value{
returns a matrix that is A rows by T columns.
}
\description{
This is a ridiculously simple function to iterate some age structured survival
that assumes a random input to age-0 each year.  It also inserts a little
extra randomness in for good measure.
}
