% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ckmr_pictorial.R
\name{make_survival_segpoints}
\alias{make_survival_segpoints}
\title{update an Es tibble to have what it needs to draw survival arrows}
\usage{
make_survival_segpoints(E, P)
}
\arguments{
\item{E}{the tibble of Ns and ERROs, etc.}

\item{P}{a tibble with t1, a1, t2, a2 giving the start and end
points of the segments}
}
\description{
This is an updated version of cum_surv_probs that lets the user degine
start and end points for any of the survival segments.
}
\details{
The function checks to make sure they are valid (i.e. that
they are within the bounds and have a slope of 1).
}
\examples{

pts <- tibble::tibble(
 t1 = c(5, 5),
 a1 = c(1, 2),
 t2 = c(11, 11),
 a2 = c(7, 8)
)


}
