% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ckmr_pictorial.R
\name{plot_ckmr_pictorial}
\alias{plot_ckmr_pictorial}
\title{Given an Eft-like tibble, make a plot displaying CKMR-related elements}
\usage{
plot_ckmr_pictorial(
  E,
  cell_fill_var = "Eft",
  cell_str_var = "Ef_str",
  min_fill = NA,
  max_fill = NA,
  cell_mask = NULL,
  survival_endpoints = NULL,
  circ_cells = NULL
)
}
\arguments{
\item{E}{a tibble like Es}

\item{cell_fill_var}{the variable to fill colors by.  Usually will be Emt or Eft
of Nt, etc.}

\item{cell_str_var}{the variable to use for the strings printed in each cell}

\item{min_fill}{min value in the fill scale}

\item{max_fill}{max value in the fill scale}

\item{cell_mask}{a tibble with columns t and a giving values of t and a to
have filled with colors (the rest will be clear/white). Values not occurring
with E will be ignored.}

\item{survival_endpoints}{a tibble with t1, a1, t2, a2 giving the start and
endpoints of the desired survival lines.}

\item{circ_cells}{a tibble with columns t and a giving the cells that should
have the little sampling circle in them}
}
\description{
Mostly this is to do a function that lets us quickly plot things with
only certain cells filled, etc.
}
\examples{
t <- 3:6
cm <- tibble::tibble(t = t, a = t+1)
cm2 <- tidyr::expand_grid(t = c(4, 7), a = 1:8)

}
