% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ckmr_pictorial.R
\name{plot_ckmr_pictorial}
\alias{plot_ckmr_pictorial}
\title{Given an Eft-like tibble, make a plot displaying CKMR-related elements}
\usage{
plot_ckmr_pictorial(
  E,
  cell_fill_var = "Eft",
  cell_str_var = "Ef_str",
  min_fill = NA,
  max_fill = NA,
  cell_mask = NULL,
  survival_endpoints = NULL,
  circ_cells = NULL,
  dcirc_cells = NULL,
  dcirc_small = 1.5,
  dcirc_big = 3,
  sample_squiggle = NULL,
  samp_squig_ystart = -0.1,
  samp_squig_fract = 0.3,
  aas_text_size = 2.5,
  aas_num_size = 2.5
)
}
\arguments{
\item{E}{a tibble like Es}

\item{cell_fill_var}{the variable to fill colors by.  Usually will be Emt or Eft
of Nt, etc.}

\item{cell_str_var}{the variable to use for the strings printed in each cell}

\item{min_fill}{min value in the fill scale}

\item{max_fill}{max value in the fill scale}

\item{cell_mask}{a tibble with columns t and a giving values of t and a to
have filled with colors (the rest will be clear/white). Values not occurring
with E will be ignored.}

\item{survival_endpoints}{a tibble with t1, a1, t2, a2 giving the start and
endpoints of the desired survival lines.}

\item{circ_cells}{a tibble with columns t and a giving the cells that should
have the little sampling circle in them}

\item{dcirc_cells}{a tibble with columns t and a giving the cells that should
have the little double-sampling circle (that denotes sampling a pair from the offspring) in them}

\item{dcirc_small}{Size of the small circle inside the dcirc.  Default = 1.5}

\item{dcirc_big}{Size of the big circle inside the dcirc.  Default = 3.0}

\item{sample_squiggle}{a tibble in which each row represents a single sampled individual.
The columns are  b (the birth year), ts (the sampling year), col (the color you want the
squiggle to be as a string)}

\item{samp_squig_ystart}{the y-value at which we place the newborn samples for the squiggles.}

\item{samp_squig_fract}{the fraction of vertical space that a year takes up in the sample squiggle
area.}

\item{aas_text_size}{the size of the "age at sampling" text if sample squiggle is not null.}

\item{aas_num_size}{the size of the age at sampling numbers if sample squiggle is not null.}
}
\description{
Mostly this is to do a function that lets us quickly plot things with
only certain cells filled, etc.
}
\examples{
vitals <- example_vitals()
set.seed(5)
E <- create_Es_from_vitals(vitals = vitals, T = 15)
plot_ckmr_pictorial(E, cell_str_var = "N")

ss <- tibble(b = c(3, 5), ts = c(6, 12), col = c("red", "blue"))

plot_ckmr_pictorial(E, cell_str_var = "N", sample_squiggle = ss)
}
